libtool: support file_magic with /path/to/dep.lib
Some platforms use same file name extension for both static and import
libraries, and an argument "import.lib" is valid both with or without a
path component when creating a shared library, and should not be added
to the static archive when it is an import library.

* build-aux/ltmain.in (func_mode_link): Support deplibs_check_method
being "file_magic" for a linker argument *.$libext, searching libpath if
necessary, to detect import libs to be handled like a -lNAME argument.