libtool: first import sym may be beyond 100th line
When using some nm to identify an import lib along import symbols,
limiting the search to the first 100 lines is insufficient.  For
example, in the MS Windows 10 SDK, for ws2_32.lib the first import
symbol is listed at line 165, and for ksuser.lib this is line 1688.

* build-aux/ltmain.in (func_win32_libid): Drop line number limit when
using some POSIX nm to query for the first import symbol listed.